# docs/api/swagger.yaml
openapi: 3.0.0
info:
  version: 0.0.1
  title: RIDI Pay API
  description: RIDI Pay API
servers:
  - url: https://api.pay.local.ridi.io
    description: Local Server
  - url: https://api.pay.ridi.com
    description: Production Server
...
paths:
...
  /users/{u_id}:
    get:
      summary: RIDI Pay 유저 정보 조회
      tags:
        - Private API
      parameters:
        - $ref: '#/components/parameters/Uid'
      responses:
        200:
          $ref: '#/components/responses/getUserResponse'
        401:
          $ref: '#/components/responses/Unauthorized'
        500:
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: 서점 회원 탈퇴 시, RIDI Pay 탈퇴 처리
      tags:
        - Private API (Only First-party)
      parameters:
        - $ref: '#/components/parameters/Uid'
      responses:
        200:
          $ref: '#/components/responses/Success'
        401:
          $ref: '#/components/responses/Unauthorized'
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /users/{u_id}/onetouch:
    put:
      summary: 원터치 결제 이용 여부 변경
      tags:
        - Private API
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                enable_onetouch_pay:
                  type: boolean
                  description: 원터치 결제 이용 여부
                  example: true
      parameters:
        - $ref: '#/components/parameters/Uid'
      responses:
        200:
          description: 'OK'
          content:
            application/json:
              schema:
                type: object
                properties:
                  is_using_onetouch_pay:
                    type: boolean
                    description: 원터치 결제 이용 여부
                    example: true
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
  /users/{u_id}/password/validate:
    post:
      summary: 입력한 계정 비밀번호 검증
      tags:
        - Private API
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
      parameters:
        - $ref: '#/components/parameters/Uid'
      responses:
        200:
          $ref: '#/components/responses/Success'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
  /users/{u_id}/pin:
    put:
      summary: 결제 비밀번호 변경
      tags:
        - Private API
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                pin:
                  type: string
                  description: 결제 비밀번호
                  example: 123456
      parameters:
        - $ref: '#/components/parameters/Uid'
      responses:
        200:
          $ref: '#/components/responses/Success'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
  /users/{u_id}/pin/validate:
    post:
      summary: 입력한 결제 비밀번호 검증
      tags:
        - Private API
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                pin:
                  type: string
      parameters:
        - $ref: '#/components/parameters/Uid'
      responses:
        200:
          $ref: '#/components/responses/Success'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServerError'
  /users/{u_idx}/payment-methods:
    get:
      tags:
        - Private API (Only First-party)
      summary: 등록된 결제 수단 목록 조회
      parameters:
        - $ref: '#/components/parameters/Uidx'
      responses:
        200:
          $ref: '#/components/responses/getPaymentMethodsResponse'
        401:
          $ref: '#/components/responses/Unauthorized'
        500:
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
    PaymentMethods:
      type: object
      properties:
        cards:
          type: array
          description: 등록된 카드
          items:
            $ref: '#/components/schemas/RegisteredCard'
    RegisteredCard:
      type: object
      properties:
        payment_method_id:
          type: string
          description: 결제 수단 ID
          example: 550E8400-E29B-41D4-A716-446655440000
        iin:
          type: string
          description: 발급자 식별 번호(카드 번호 앞 6자리)
          example: 541654
        issuer_name:
          type: string
          description: 카드 발급사명
          example: 신한
        color:
          type: string
          description: 카드 플레이트 색상
          example: 00FF00
        logo_image_url:
          type: string
          description: 카드 발급사 로고 이미지 URL
        subscriptions:
          type: array
          description: 연결된 정기 결제 서비스명
          items:
            type: string
          example: [리디캐시 자동충전, 리디셀렉트]
      required:
        - payment_method_id
        - iin
        - issuer_name
        - color
        - logo_image_url
        - subscriptions
    RegisteredCardSimple:
      type: object
      properties:
        payment_method_id:
          type: string
          description: 결제 수단 ID
          example: 550E8400-E29B-41D4-A716-446655440000
        iin:
          type: string
          description: 발급자 식별 번호(카드 번호 앞 6자리)
          example: 541654
        issuer_name:
          type: string
          description: 카드 발급사명
          example: 신한
      required:
        - payment_method_id
        - iin
        - issuer_name
...
    FailureResponse:
      type: object
      properties:
        code:
          type: string
          description: 에러 코드
        message:
          type: string
          description: 사람이 읽을 수 있는 에러 메시지
      required:
        - code
        - message
  parameters:
...
    Uid:
      name: u_id
      in: path
      required: true
      description: Ridibooks 유저 ID
      schema:
        type: string
    Uidx:
      name: u_idx
      in: path
      required: true
      description: Ridibooks 유저 고유 번호(IDX)
      schema:
        type: integer
  responses:
...
    Success:
      description: OK(200)
      content:
        application/json:
          schema:
            type: object
    BadRequest:
      description: Bad Request(400)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/FailureResponse'
    Unauthorized:
      description: Unauthorized(401)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/FailureResponse'
    Forbidden:
      description: Forbidden(403)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/FailureResponse'
    NotFound:
      description: Not Found(404)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/FailureResponse'
    InternalServerError:
      description: Internal Server Error(500)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/FailureResponse'
...
    getUserResponse:
      description: 'OK'
      content:
        application/json:
          schema:
            type: object
            properties:
              payment_methods:
                type: object
                description: 등록된 결제 수단 목록
                $ref: '#/components/schemas/PaymentMethods'
              has_pin:
                type: boolean
                description: 결제 비밀번호 소유 여부
                example: true
              is_using_onetouch_pay:
                type: boolean
                description: 원터치 결제 이용 여부
                example: true
            required:
              - payment_methods
              - has_pin
              - is_using_onetouch_pay
    getPaymentMethodsResponse:
      description: 'OK'
      content:
        application/json:
          schema:
            type: object
            properties:
              cards:
                type: array
                description: 등록된 카드
                items:
                  $ref: '#/components/schemas/RegisteredCardSimple'